System message to each game on player join and leave
