+++ backend/views/home.pug
@@ -47,3 +47,6 @@ block content
           div(id = 'messages')
               label
                   input(id='chatMessage',  name='chat-message',placeholder='Message...')
+
+  script(src='/scripts/gameTable.js')
+  script(src='/scripts/homeChat.bundle.js', defer)
