+++ backend/views/layout.pug
@@ -2,8 +2,6 @@ html
   head
     title #{title}
     link(rel='stylesheet' href='/stylesheets/home.css')
-    script(src='scripts/gameTable.js')
-    script(src='scripts/bundle.js', defer)
 
   body
     header
